Define alwaysinline functions at the end of
DtoDeclareFunction even if they are not externallyAvailable.
This way, those functions are inlined even if they are
defined in a module that we are not currently generating code
for, but was also given on the command line,
and the -singleobj flag was not used. Also made alwaysInline
functions have private linkage, so that they don't appear in
the object file. This avoids linker errors in the case
described above.